test math 71 with only new additions

+++ src/checkout-gitshow-hashing/main.py
@@ -18,21 +18,27 @@ class MainProcessor:
         self.csv_writer = CSVWriter(output_dir)
 
     def process(self, project, bug_num):
-        project_dir = self.project_checkout.checkout(project, bug_num)
-        diff = GitDiffParser.get_diff(project_dir)
-
-        # Save diff to folder "/root/data/Defects4J/project_diffs/{project}_{bug_num}.txt"
-        diff_file_path = f"/root/data/Defects4J/project_diffs/{project}_{bug_num}.txt"
+        diff_file_path = f"/root/data/Defects4J/project_diffs/{project}_{bug_num}.diff"
         os.makedirs(os.path.dirname(diff_file_path), exist_ok=True)
-        with open(diff_file_path, 'w') as diff_file:
-            diff_file.write(diff)
-            
-        # methods = MethodExtractor.extract_methods(diff)
-        # base_path = f'/tmp/repos/{project}_{bug_num}/'
-        # method_implementations = MethodExtractor.extract_method_implementations(diff, methods, base_path)
+        
 
-        # # for method_name, implementation in method_implementations.items():
-        # #     print(f"Method: {method_name}\nImplementation:\n{implementation}")
+        if not os.path.exists(diff_file_path):
+            project_dir = self.project_checkout.checkout(project, bug_num)
+            diff = GitDiffParser.get_diff(project_dir)
+
+            # Save diff to folder "/root/data/Defects4J/project_diffs/{project}_{bug_num}.txt"
+            with open(diff_file_path, 'w') as diff_file:
+                diff_file.write(diff)
+        else:
+            with open(diff_file_path, 'r') as diff_file:
+                diff = diff_file.read()
+        
+        methods = MethodExtractor.extract_methods(diff)
+        base_path = f'/tmp/repos/{project}_{bug_num}/'
+        method_implementations = MethodExtractor.extract_method_implementations(diff, methods, base_path)
+
+        for method_name, implementation in method_implementations.items():
+            print(f"Method: {method_name}\nImplementation:\n{implementation}")
 
         # method_triples = []
         # for file_path, method_name_line_pairs in methods.items():
